apps/oracle-data-migration-harness/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oracle-data-migration-harness"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "oracledb>=2.4",
    "langchain>=0.3",
    "langchain-oracledb>=0.1",
    "langchain-openai>=0.2",
    "pymongo>=4.10",
    "python-dotenv>=1.0",
    "pydantic>=2.9",
    "sse-starlette>=2.1",
    "tenacity>=9.0",
    "openai>=1.51",
    "numpy>=1.26",
    "sentence-transformers>=3.0",
    "oci>=2.170",
]

[project.optional-dependencies]
dev = ["pytest>=8.3", "pytest-asyncio>=0.24", "httpx>=0.27"]

[tool.setuptools.packages.find]
where = ["."]
include = ["data_migration_harness*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: tests that require live Oracle Database and/or MongoDB services",
]
addopts = "-m 'not integration'"
